+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### Added
 - **Control Flow Graphs for managed and native methods**
+- Option to disable underlining of managed assemblies
 
 ## [0.4.1] - 2020-07-24
 ### Fixed